moved games out into separate files
They're imported back into the book with a fragile-ish import
mechanism.

This move makes the games easier to edit, separate from their
LaTeX chapters. The commentary about the games is now a bit
separated, but that's probably OK.

I also took the opportunity to make sure all of the games
(except the adventure game) run on latest nom.

I'm really dissatisfied with the way the winning condition
check looks in connect-4. It needs re-working.